forgot to do a little error checking, put it in

+++ refinery/core/management/commands/init_refinery.py
@@ -25,9 +25,11 @@ class Command(BaseCommand):
         # 1. test if admin user exists
          
         # 2. set up the site name
+        if Group.objects.filter(domain__exact=settings.REFINERY_BASE_URL).count() > 0:
+            raise CommandError("This URL already exists in the database")
         s = Site(name=settings.REFINERY_INSTANCE_NAME, domain=settings.REFINERY_BASE_URL)
         s.save()
-        #remove dummy one if it exists
+        #remove dummy site if it exists
         try:
             s = Site.objects.get(domain="example.com")
             s.delete()
